requirements.txt
onnxruntime>=1.10.0
pillow>=8.0.0
rembg>=2.0.0
flask>=2.0.0
waitress>=2.1.2
